Modified state definitions, added eval board FM

+++ RTC.c
@@ -44,7 +44,7 @@ time_t getTime(void)
 *	\param decimal The decimal form of a number
 *	\return BCD form of decimal number
 */
-int bcdEncode(uint8_t decimal)
+uint8_t bcdEncode(uint8_t decimal)
 {
 	return ((decimal / 10)<<4) | (decimal % 10);
